[TASK] swift mailer update to version 4.1.3
Resolves: #32062
Releases: 4.7

+++ typo3/contrib/swiftmailer/classes/Swift/KeyCache/DiskKeyCache.php
@@ -8,12 +8,6 @@
  * file that was distributed with this source code.
  */
 
-//@require 'Swift/KeyCache.php';
-//@require 'Swift/KeyCacheInputStream.php';
-//@require 'Swift/InputByteStream.php';
-//@require 'Swift/OutputByteStrean.php';
-//@require 'Swift/SwiftException.php';
-//@require 'Swift/IoException.php';
 
 /**
  * A KeyCache which streams to and from disk.
@@ -29,6 +23,9 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
 
   /** Signal to place pointer at end of file */
   const POSITION_END = 1;
+  
+  /** Signal to leave pointer in whatever position it currently is */
+  const POSITION_CURRENT = 2;
 
   /**
    * An InputStream for cloning.
@@ -68,7 +65,11 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
   {
     $this->_stream = $stream;
     $this->_path = $path;
-    $this->_quotes = get_magic_quotes_runtime();
+
+    if (function_exists('get_magic_quotes_runtime') && @get_magic_quotes_runtime() == 1)
+    {
+      $this->_quotes = true;
+    }
   }
 
   /**
@@ -99,6 +100,7 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
         break;
     }
     fwrite($fp, $string);
+    $this->_freeHandle($nsKey, $itemKey);
   }
 
   /**
@@ -133,6 +135,7 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
     {
       fwrite($fp, $bytes);
     }
+    $this->_freeHandle($nsKey, $itemKey);
   }
 
   /**
@@ -171,7 +174,7 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
       $fp = $this->_getHandle($nsKey, $itemKey, self::POSITION_START);
       if ($this->_quotes)
       {
-        set_magic_quotes_runtime(0);
+        ini_set('magic_quotes_runtime', 0);
       }
       $str = '';
       while (!feof($fp) && false !== $bytes = fread($fp, 8192))
@@ -180,8 +183,9 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
       }
       if ($this->_quotes)
       {
-        set_magic_quotes_runtime(1);
+        ini_set('magic_quotes_runtime', 1);
       }
+      $this->_freeHandle($nsKey, $itemKey);
       return $str;
     }
   }
@@ -199,7 +203,7 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
       $fp = $this->_getHandle($nsKey, $itemKey, self::POSITION_START);
       if ($this->_quotes)
       {
-        set_magic_quotes_runtime(0);
+        ini_set('magic_quotes_runtime', 0);
       }
       while (!feof($fp) && false !== $bytes = fread($fp, 8192))
       {
@@ -207,8 +211,9 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
       }
       if ($this->_quotes)
       {
-        set_magic_quotes_runtime(1);
+        ini_set('magic_quotes_runtime', 1);
       }
+      $this->_freeHandle($nsKey, $itemKey);
     }
   }
 
@@ -232,11 +237,9 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
   {
     if ($this->hasKey($nsKey, $itemKey))
     {
-      $fp = $this->_getHandle($nsKey, $itemKey, self::POSITION_END);
-      fclose($fp);
+      $this->_freeHandle($nsKey, $itemKey);
       unlink($this->_path . '/' . $nsKey . '/' . $itemKey);
     }
-    unset($this->_keys[$nsKey][$itemKey]);
   }
 
   /**
@@ -251,7 +254,10 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
       {
         $this->clearKey($nsKey, $itemKey);
       }
-      rmdir($this->_path . '/' . $nsKey);
+      if (is_dir($this->_path . '/' . $nsKey))
+      {
+        rmdir($this->_path . '/' . $nsKey);
+      }
       unset($this->_keys[$nsKey]);
     }
   }
@@ -286,20 +292,31 @@ class Swift_KeyCache_DiskKeyCache implements Swift_KeyCache
    */
   private function _getHandle($nsKey, $itemKey, $position)
   {
-    if (!isset($this->_keys[$nsKey]) || !array_key_exists($itemKey, $this->_keys[$nsKey]))
+    if (!isset($this->_keys[$nsKey][$itemKey]))
     {
-      $fp = fopen($this->_path . '/' . $nsKey . '/' . $itemKey, 'w+b');
+      $openMode = $this->hasKey($nsKey, $itemKey)
+        ? 'r+b'
+        : 'w+b'
+        ;
+      $fp = fopen($this->_path . '/' . $nsKey . '/' . $itemKey, $openMode);
       $this->_keys[$nsKey][$itemKey] = $fp;
     }
     if (self::POSITION_START == $position)
     {
       fseek($this->_keys[$nsKey][$itemKey], 0, SEEK_SET);
     }
-    else
+    elseif (self::POSITION_END == $position)
     {
       fseek($this->_keys[$nsKey][$itemKey], 0, SEEK_END);
     }
     return $this->_keys[$nsKey][$itemKey];
+  }
+  
+  private function _freeHandle($nsKey, $itemKey)
+  {
+    $fp = $this->_getHandle($nsKey, $itemKey, self::POSITION_CURRENT);
+    fclose($fp);
+    $this->_keys[$nsKey][$itemKey] = null;
   }
 
   /**
